lekce_9/mcp_server/pyproject.toml
[project]
name = "wazuh-mcp-server"
version = "0.1.0"
description = "MCP server for querying Wazuh incidents from OpenSearch"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.3.0",
    "opensearch-py>=2.4.0",
    "uvicorn>=0.27.0",
    "starlette>=0.36.0",
    "python-dotenv>=1.0.0",
    "ddgs>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["server.py", "tools"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
